New line

+++ Ex12a/src/ru/itis/Main.java
@@ -4,7 +4,7 @@ public class Main {
 
     public static void main(String[] args) {
 	Scanner in=new Scanner(System.in);
-	int x=in.nextInt();
+	double x=in.nextInt();
 	int n=1;
 	double eps=0.0001,s=0,t=1;
 	while(Math.abs(t)>eps){
